+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/opmodes/tuning/AutomaticFeedforwardTuner.kt
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.autonomous.drive.opmode
+package org.firstinspires.ftc.teamcode.autonomous.opmodes.tuning
 
 import com.acmerobotics.dashboard.FtcDashboard
 import com.acmerobotics.dashboard.config.Config
