+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/opmodes/tuning/BackAndForth.kt
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.autonomous.drive.opmode
+package org.firstinspires.ftc.teamcode.autonomous.opmodes.tuning
 
 import com.acmerobotics.dashboard.config.Config
 import com.acmerobotics.roadrunner.geometry.Pose2d
